Only show LFS misconfiguration banner when objects exist in project repo
When we previously enabled the lfs_misconfiguration_banner flag we saw
reports of banners being incorrectly displayed. We discovered that
LfsObject records do not necessarily belong to the project's repository
as wikis and design repositories also use LFS to store files.

After this change we now check the LfsObjectsProject records instead and
scope to records that have a repository_type of project.

This also adds a project actor to the feature flag.

Changelog: fixed