Fix is_blocked_by for epic work item
Fixes a bug where setting `is_blocked_by` as the link_type when the
source is not an epic didn't work.

We can't sync issue<>epic related links to legacy epics. The existing
check only ensured that the `target` is an epic. However, when
`is_blocked_by` is used, the target and the source switch.

The same bug would have happened when we'd set an issue work item
(as the target) to be related to an epic work item.

Changelog: fixed
EE: true