Ensure access_level is present before comparison
This commit addresses an ArgumentError exception in Sentry caused by
comparing an Integer with nil in the Member#higher_access_level_than_group
validation:

```
ArgumentError
comparison of Integer with nil failed (ArgumentError)

if highest_group_member && highest_group_member.access_level > access_level
```

The validation assumes access_level is present, but this isn't
guaranteed. Although we have an access_level_inclusion validation to
ensure access_level is not nil, the order of validations in Rails is
not predetermined.

To resolve this, we modify the validation to work even when some
attributes are nil, ensuring robustness regardless of validation order.

Changelog: fixed