Remove fixed specs from quarantine
This should be one of the last times this is necessary, as
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/174250 and
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/174365 will ensure
that merge requests must remove specs from quarantine when they would
otherwise pass.

The only remaining way that something might slip through is if a
security merge request fixes a spec, but that is very unlikely to
happen.