Namespace api find un-replicated newly created trial group by id
Gitlab.com trial has a `create group flow`. In this flow, the Customer
create a new group, and passing the `group.id` to CustomersDot API
to create the trial. CustomersDot API will find the namespace data by
calling GitLab namespace API, using the group ID. Sometimes the Gitlab
database has a delay to replicate from primary DB to replicate DB.
In this case, when CustomersDot find the namespace data using Gitlab
namespace API, there is a chance the newly created group is not found
in replication DB in Gitlab. This causes 404 not found error.

This MR stick the namespace ID when the new group is created in the
`create group trial flow`. Then in Gitlab namespace API, it query the
namespace from up-to-date replica DB(or primary DB). This ensures the
newly created group be found correctly.

Changelog: fixed
EE: true