Rename existing jobs as predictive
They are predictive in the sense that they only check specs which are
affected by the merge request's changes, relying on Jest's
`--findRelatedTests` option. This is exactly how other predictive Jest
jobs work, so it makes sense to name them similarly.

This also changes the rules of the jobs so that they run when the other
Jest predictive jobs would run, i.e., approximately in tier 1 pipelines.
This means authors should get feedback sooner if they need to remove
specs from quarantine.