Continue to show roadmap filters bar when no results to access settings
When there are no epics results for the roadmap, continue to show the
fitlers bar, but hide the search and filter input. This allows the
user to continue to access the Settings dropdown, which may be
necessary if, for example, they are seeing the empty state because they
have selected "show closed epics" but there are none.

Changelog: fixed
EE: true