Fix Psych::DisallowedClass in WebHookLog
As our WebHook log request data is serialized as YAML, if ever a data
builder generates a payload with values that are not permitted by our
safe YAML rules, we could throw a `Psych::DisallowedClass` when
attempting to save the log to PostgreSQL. Normally, because the payload
would be sent to Redis first before being picked up by a worker, the
payload would be JSONified through this process. But during a webhook
test, the webhook is executed in request. In this situation we could see
exceptions due to the unsafe YAML value.

This change makes the payload the same as if it had been sent to Redis
first, as it will be if the service is called in-request.

https://gitlab.com/gitlab-org/gitlab/-/issues/386981

Changelog: fixed