Restrict access for duo workflow
Prevents access to projects and groups to "Duo Workflow user". Because there is
currently no real "Duo Workflow user" (Duo Workflow uses user's
account), the only difference is that it uses a special token for
authenticating its requests (this token has "ai_workflows" scope).

If this token is used for authenticating a request and it attempts to
access a project or group which has Duo features disabled, it prevents
access to this resource.