Add budget into Auto Resolve process
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/174454, we
added a limit to Auto Resolve. This introduced a minor edge case where
the limit could be breached because the next run of the Auto Resolve
process didn't consider any previously-completed work.

In the worst case, the current implementation would allow
(2 * `limit` - 1) records to be resolved per run, instead of `limit`.

This MR introduces a "budget" for the Auto Resolve process where it
instructs the auto_resolve_service to attempt to resolve a maximum
number of records ("budget") within the batch. The budget is computed
as the difference between the `limit` and the number of successfully
Auto Resovled vulnerabilities in this run.

Changelog: fixed
EE: true