Reset unconfirmed_email if it exists on /users/identity_verification/verify_email_code
Users who attempted to register prior to the
email_confirmation_setting set to hard in Gitlab.com
would have their primary email changed to the email set
in the unconfirmed_email attribute, if present. This is
an issue because the email confirmation code is sent to
the user's primary email address and would result in a user
confirming the wrong email. This MR adds a check to remove
the unconfirmed_email address of a user before confirmation
in the registrations identity verification controller.

Changelog: fixed
EE: true