Fix race conditions in SourceUserMapper
We see `ActiveRecord::RecordInvalid: Validation failed: Source user
identifier has already been taken` errors from
`#create_source_user_mapping` despite being in an exclusive lease that
has all properties scopes to the uniqueness of `source_user_identifier`.

The theory is there can be a race condition where:

- while finding a source user in
  https://gitlab.com/gitlab-org/gitlab/-/blob/f8c2e89b4e3d594c62c866be79d327bcbfc8bab2/lib/gitlab/import/source_user_mapper.rb#L47
  the source user didn't yet exist (but potentially a lease had been
  taken by another thread)
- when encountering the lease in
  https://gitlab.com/gitlab-org/gitlab/-/blob/f8c2e89b4e3d594c62c866be79d327bcbfc8bab2/lib/gitlab/import/source_user_mapper.rb#L71
  the source user has been written and the lease has been released, so
  we obtain the lease. We haven't retried, so we don't attempt to find
  the source user that was created already.
- we save
  https://gitlab.com/gitlab-org/gitlab/-/blob/f8c2e89b4e3d594c62c866be79d327bcbfc8bab2/lib/gitlab/import/source_user_mapper.rb#L95
  and find out the record is not unique

This change does not rely on `retried` as the indicator that we should
find the source user within the lease, but instead if we experience a
`DuplicatedSourceUserError`.

https://gitlab.com/gitlab-org/gitlab/-/issues/502708#note_2223402124