Move nuget symbol files creation outside uploading package transaction
nuget symbols creation is wrapped in the same database transaction as the main package upload. That means an error during nuget symbols creation can abort the transaction and then the whole process with be aborted. We need to avoid this by moving the symbols creation outside the transaction.

Changelog: fixed