Fix adding child items to work item epics
Fix permissions when setting a work item as parent

When using the WorkItems::ParentLink::CreateService to set the parent
of a work item, we checked if the user has the permission to admin the
parent.

This is however a different permission than we had on the legacy epic
experience, where we only ensured that the user has the permission to
read the parent.

This affects now users who are only a member of the project but not of
the group. They would not be allowed to set the parent epic of an issue
anymore, even though they are able to read the epic.