Improve performance of feature checks with actor
On GitLab.com, we have a `gitlab_team_members` feature group that is
checked whenever we check if a feature flag is enabled for an actor.

The list of user IDs is cached using `ProcessMemoryCache` but
deserializing the data from cache every time a feature flag is checked
can get expensive.

We cache the data in RequestStore so that we only need to deserialize
this once per request.

Changelog: performance
EE: true