Fix CI_PAGES_DOMAIN and CI_PAGES_URL for parallel deployments
This change moves the creation of the CI variables
CI_PAGES_DOMAIN and CI_PAGES_URL to models/build.rb and uses
Pages::UrlBuilder to propagate its values. This ensures that
a `path_prefix` provided in the job config is reflected
in the value.

It also introduces a declarative refactor of the
Pages::UrlBuilder, hence the changelog

Changelog: changed
Signed-off-by: janis <jaltherr@gitlab.com>