Additional updates to PAT buffered expiration flow
Fix the placeholder text for the maximum length of the PAT token in
the documentation string for the EE feature that sets the maximum
allowable days for a token.

Add in a "max" attribute that limits the value here to the maximum
value controlled by the buffer_token_expiration_limit.

Introduce a class method for the checking of the timeout and consolidate
the other places where this check is used to use this class method.

See issue: https://gitlab.com/gitlab-org/gitlab/-/issues/508151

Changelog: fixed