Make discussions API work with quick actions-only notes
This is a follow-up to
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/173055.

Previously quick actions were not handled properly in the commits and
discussions API for the following API routes:

* POST
  /projects/:id/<noteable>/<noteable_id>/discussions/:discussion_id/notes
* PUT
  /projects/:id/issues/:issue_iid/discussions/:discussion_id/notes/:note_id

Users would see a 400 Bad Request even if the quick action were
initiated. This commit changes that to respond with a 202 upon
success.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/505100

Changelog: fixed