Deprecate :labels negatable filter value
This mismatch caused negated label filters to not work.

We can't change any of the other values around this field
because they're also used by the Issues list, so we need to
update the name of this negatable field.

Luckily, `labels` is the deprecated name, so we shouldn't have
any problem doing this.

Changelog: fixed