Run release-environments pipeline from the same branch
Previously, the release-environments pipeline is always triggered from
`master` pipeline so that we don't need to backport pipeline changes to
stable branches whenever a change is made. However, this strategy
doesn't fit anymore:

- The pipeline is quite stable now, so we don't make changes often.
- The mismatch between the commit ref using for the upstream and
downstream pipelines sometimes leads to unexpected behaviors (e.g.
a QA image build in `master` may not work with the QA setup in the
stable branches)

Ref: https://gitlab.com/gitlab-org/release/tasks/-/issues/15464