Ensure p element inside details summary behaves properly
When an inline code element in rendered markdown has a long string with
no whitespace in it, today that code element will expand horizontally
outside of its parent boundaries if it is inside the default `<p>`
element rendered inside a `<summary>` when the summary contains
markdown. This small change to the layout of the `<p>` element inside a
summary fixes this.

Changelog: fixed