Handle null users in GitHub user mapping
When we imported the rails repository from GitHub we saw some
`NoMethodError: undefined method 'id' for nil:NilClass` in places where
we push placeholder references
https://gitlab.com/gitlab-org/gitlab/-/issues/502708#note_2223372306.

It happens because sometimes the GitHub API doesn't return a user when
we expect it to - perhaps this is when the GitHub user has been blocked
or deleted.

For example, the first `NoMethodError` in the table, we expect a
`request_reviewer` for a `review_requested` event,
https://gitlab.com/gitlab-org/gitlab/-/blob/3ff8e1872928fcb4ddabff5d7d0a9b2f6a440d3b/lib/gitlab/github_import/importer/events/changed_reviewer.rb#L41
however the API response
https://api.github.com/repos/rails/rails/issues/40864/timeline for event
with ID `4127612511` has no request_reviewer (see the payload on the
confidential issue
https://gitlab.com/gitlab-org/gitlab/-/issues/502708#note_2223434852.

Through the `UserFinder`, we import data that has no author to GitLab
associating it with `User.ghost`.

In these situations, we shouldn't push any placeholder references.
The user we mapped to is `User.ghost`, so we should consider that they
have accepted the contribution mapping.

This applies to every call to `#push_with_record`,
`#push_with_composite_key`, and `#push_refs_with_ids`.

https://gitlab.com/gitlab-org/gitlab/-/issues/505756