Implement Duo Code Review Chat
We are using Duo Chat for the chat functionality part of Duo
Code Review.

This can be triggered by tagging GitLab Duo in a comment in a
diff discussion the bot in started.

We send most of the notes in thread as part of context we send along
in our duo chat request. We exclude notes that are not mentioning
GitLab Duo as it's not needed and may confuse the chat bot if we
include comments that aren't related to the review (e.g. other
users chiming in).

We also send the relevant lines of code as additional context.

This is behind `duo_code_review_chat` FF.