Extend widget polling to after Security::StoreFindingsService
Background context
---------------------

We had an [epic][0] to move from using report artifices for
vulnerability finding comparisons. That epic is complete, and the
pertinent information is now available in the `finding_data` `jsonb`
column on the [security_findings table][1]

The [security MR widget][2] is one feature that relied on the original
technique of using report artifacts for comparisons. As such, it can
now be migrated to using the new `finding_data` column.

This Change
-------------------

As outlined in [this thread][3], before we migrate the security MR
widget to use the database, we need to make sure the merge request
controller stays polling until the database is actually
ready. [^1]

The current behavior correlates to ending the polling at step 3 in the
`Scan runs in a pipeline for a non-default branch` [flow][7]. We need
to move the polling to last until after step 7 of the same flow (after
the [`Security::StoreFindingsService`][8] has completed its work

The Redis caching code
--------------------------

The security MR widget must poll until all scans are stored in the
database. We determine if the scans are fully ingested by running the
query introduced to `CompareSecurityReportsService` in this change

This query is somewhat expensive, especially in the context of this
being used in a polling endpoint.

All of this redis caching logic exists to reduce load on the database
during that polling. For the duration of
SECURITY_MR_WIDGET_POLLING_CACHE_TTL, we skip querying the database
and rely on checking the existence of this redis keys instead to know
whether we should be polling or not

the logic is:

1. In the pipeline's `before_transition` hook, we set a key that will
   make the reports endpoint return a parsing response (and therefore
   poll)

2. In the `StoreGroupedScansService`, after a given report type is
   finished being stored, a separate redis key is set. That key is used
   to allow the reports endpoint to call the finder and return the
   results while bypassing the query I am introducing in
   `CompareSecurityReportsService`

3. all of the redis keys have a 2 hour TTL, after which the new query
   in CompareSecurityReportsService will be used

[^1]: This requests [currently polls][4] while `pipeline&.active?`. It
      also polls while the report has a status of `parsing` [here][5],
      which is set in the MR model [here][6]

[0]:https://gitlab.com/groups/gitlab-org/-/epics/8341
[1]:https://gitlab.com/gitlab-org/gitlab/-/blob/ef9cee18a0a4c75a0e3c691297b8d964315829cd/db/structure.sql#L1218
[2]:https://docs.gitlab.com/ee/user/application_security/#merge-request
[3]:https://gitlab.com/gitlab-org/gitlab/-/issues/390185#note_1374606930
[4]:https://gitlab.com/gitlab-org/gitlab/-/blob/55151efbb58d45587fc44317f9fe95307fa1add3/app/controllers/projects/merge_requests_controller.rb#L588-592
[5]:https://gitlab.com/gitlab-org/gitlab/-/blob/master/app/controllers/projects/merge_requests_controller.rb#L594-597
[6]:https://gitlab.com/gitlab-org/gitlab/blob/master/app/models/merge_request.rb#L1845-1852
[7]:https://gitlab.com/gitlab-org/gitlab/-/blob/master/doc/development/sec/security_report_ingestion_overview.md#vulnerability-creation-from-security-reports
[8]:https://gitlab.com/gitlab-org/gitlab/-/blob/master/ee/app/services/security/store_findings_service.rb

---

related to: https://gitlab.com/gitlab-org/gitlab/-/issues/390185
resolves: https://gitlab.com/gitlab-org/gitlab/-/work_items/457851
Changelog: changed
EE: true