Step-up auth: Add feature tests for step-up auth scenarios
In the MR https://gitlab.com/gitlab-org/gitlab/-/merge_requests/171643,
we implemeted the step-up auth feature for the admin mode
with unit tests, but without feature tests

In this MR we are adding basic feature tests to cover the whole
step-up auth user flow.
It also includes more advanced feature tests
for scenarios with different initial sign-in methods.

Changelog: other