Enable closing issues automatically, cross-project
Now you can configure an issue to be closed by ANY merge
request as long as you have read access to that merge request.
Even if your issue is private and the user mergeing the merge request
doesn't have access to your issue, the issue will be closed.

Currently, the only way to configure this is through the GraphQL
mutation workItemAddClosingMergeRequest

Changelog: changed