Queue AfterImportWorker after every import
`Projects::AfterImportWorker` does some repository housekeeping after
an import. We call this when the `ProjectImportState` transitions to
finished, but previously only if the import source had an importer
class. This change will call the worker also when the import source was
`git`, which has no special importer but does has a repository after
the import has finished.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/174610#note_2254600232

Changelog: changed