Set fixed order of labels to fix flaky spec
There are tests in label_note_spec.rb that checks the sorting
of events by label title. For some reason, the `label` and `label2`
factories can be created with titles that have a different sorting.
As a result, the spec will fail randomly.

To ensure that order is fixed, we set fixed titles for the labels
used in specs.