Increase ci_max_artifact_size_lsif default limit
In order to support bigger LSIF artifacts, we are increasing the
default limit of LSIF artifacts.

This is currently increased on production in code behind the
`increase_lsif_artifacts_limit` feature flag which will be removed
after these migration run. That is needed to be done so GitLab.com
won't revert to lower limits until we run these migrations.

Changelog: changed