Refactor services to be job centric
We refactor `Ci::PlayBuildService` and
`Ci::RunScheduledBuildService` to rely solely
on the CI job and `current_user` to be passed
in as parameters rather than accepting `project`
and sometimes `current_user` separately.

This improves invariants around data integrity
which assume that the `project` and sometimes
`current_user` must match those of the `Ci::Build`.