Add feature flags for all included duo context
The new features to enable users to add contexts via a context picker
drop down have been implemented behind a single feature flag. This means
that we could only roll them out all together or not at all.

To correct this we add in feature flags for each individual context type
that will be included in Duo Chat and update the GraphQL query to check
if the feature flag for that type context is enabled for the user.

EE: true
Changelog: changed