Improve move and clone work item specs performance
Instead of running the move or clone service for each widget
we can run it once with all widgets and check that the output
is the expected one for each widget data. This reduces the time
spent on running the spec as we run it once vs 8 times now and
more times as we keep adding widgets.

This also prevents issues where the order in which some widgets
are run is important.