Ignore EE specs in FOSS pipelines
This fixes an edge case when checking the quarantine of Vue 3 specs
under FOSS pipelines.

One feature of this script is that it reports specs that should be
removed from quarantine when they no longer exist. This is because
quarantining something that does not exist is pointless.

The problem is that in FOSS pipelines, the `ee/` directory is removed:
https://gitlab.com/gitlab-org/gitlab/-/blob/28a3562ebc9807bc962a7dc653be0fb118b728c8/.gitlab/ci/as-if-foss.gitlab-ci.yml#L27,
so no EE specs exist. This script should take that into account, and not
report EE specs in that case.

This came up in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/175824.