Handle errors during Duo Code Review chat
It's possible that an error occurs during DCR chat. We need to
handle it gracefully and post a reply to the same thread where
DCR chat is executed.

We're also updating the copy of error note for DCR, to mention
`problems` instead of `issues`.

This is still behind `duo_code_review_chat` FF.