Handle 404s when fetching GitHub user details
The GitHub API can return a `404` when we try to fetch the user's
details. We can see some errors when importing `rails/rails`.
Specifically for the user `Copilot`.

We already have a fallback of using the GitHub username as the source
user's name if we're unable to get the username from the API response.
https://gitlab.com/gitlab-org/gitlab/-/blob/331d2e9318220e6e229857aa4f0b7eb49e1e87ed/lib/gitlab/github_import/user_finder.rb#L133

This change does the same if we receive a `404` from the GitHub API.