Use custom ResponseBodyParser
Nokogiri is stripping reserved XML characters like `&` and `<` which we
don't want as we need to display them as is within the code suggestion.
Since LLM's response isn't a strict XML, we don't need to worry about
stripping them out and just use a custom parser.

Also, it handles a few more edge cases such as multiple code suggestions
in single comment and handling <from> <to> tags inside of the code
suggestion.