Send AutoFlow environment scoped CI variables with events
This change set introduces a boring solution for KAS to access secrets
defined by a project. The lack of a built-in secrets manager makes it
hard for KAS to access such things. This implementation uses CI
variables as a store for secrets. The special environment
`autoflow/internal-use` must be assigned to the environment variable. When
Rails sends an event to AutoFlow it attaches all CI variables with that
special environment scope to the event.

Refs https://gitlab.com/gitlab-org/gitlab/-/issues/508258