Fix backoff of disabled webhooks ramp up too fast
When a webhook is erroring it is told to back off. Once
`recent_failures` reaches 4, and there is a `disabled_until` set for
the future, the webhook will no longer trigger:
https://gitlab.com/gitlab-org/gitlab/-/blob/b9a2017b64d56cb56be6eeac19c2c146a845154c/app/models/concerns/web_hooks/auto_disabling.rb#L65

But if there are many webhooks that were triggered in a short period of
time, many/all of which fail, previously all of those failures would
individually contribute to the `backoff_count` count, and therefore
all individually cause the `disabled_until` to be exponentially set to
a time in the future, as `disabled_until` is "ramped up" based on the
`backoff_count`.

This is contrary to the expected behaviour of the feature, which is that
if a webhook receiver is experiencing a failure for, say, a minute, it
would be expected that webhook could not jump from being enabled to
being disabled for a long period of time, due to there having been a
number of webhooks triggered during the webhook receiver failure before
they started being recorded as having failed.

In this change we will only increment the `backoff_count` if the webhook
is not already recorded as being temporarily disabled. Due to the lease
and reloading of the webhook when handling errors, this logic will work
during periods of many simultaneous webhook failures.
https://gitlab.com/gitlab-org/gitlab/-/blob/a71e0a5e92cbb4f7c3968f6e8450cdbc48d62b0e/app/services/web_hooks/log_execution_service.rb#L49-50

Changelog: fixed