Use data from current tab of Vue MR list page for export to CSV
Currently, the values for the csv export path and the count of items to
be exported when viewing the Vue MR list page is loaded once on page
load, which means that only the items in the tab that is loaded
initially get exported. This MR adjusts this to use the data from the
current tab for export.

Changelog: fixed