Fix GraphQL removals being labelled as deprecations
We erroneously were describing the GraphQL removals page as being a
page of deprecations when linking to it from the main GitLab
Deprecations page.

Instead, GraphQL deprecations are listed on that main GitLab
Deprecations page, alongside other features that are deprecated.

The REST API deprecations _are_ documented separately because the REST
API versioning doesn't follow GitLab's, and so anything "deprecated" in
REST v4 would be removed in v5, but there is no timeframe for v5 at
present or ability to say what GitLab version it might happen in.

So we keep the REST API deprecations page being a separate document of
deprecations.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/175953#note_2266501626