Fix foreign key error while backfilling
The foreign key contraints on parent_id on namespaces table
has 'NOT VALID' clause.
This means there can be old recods which is not valid,
and has parent_id that does not exists in database.

This MR fixes the Background Migration to handle such records
by existing if user_id and root_namespace_id already exists.

EE: true
Changelog: fixed