Run EnsureFactoryForTable in CE to avoid RedundantCopDisableDirective
Currently, we skip the Migration/EnsureFactoryForTable cop when
running in CE. This results in a Lint/RedundantCopDisableDirective
offense when there's a
"rubocop: disable Migration/EnsureFactoryForTable" comment.

To avoid this, we now run the Migration/EnsureFactoryForTable cop
when a disable comment is present. This prevents the redundant
disable directive offense.

Changelog: other