Minor cleanup of ee/spec/requests/api/internal/base_spec.rb
Previously we could not use `let_it_be` with the service account
because the composite identity was behind a `composite_identity_auth`
license check. In tests, the default license is a starter license, so
the `composite_identity_auth` feature was not available. Since tests
are run in a transaction, the `before` block would run
`primary_user.update!(composite_identity_enforced: true)`, but this
change would be rolled back after the first example.

In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/175526 we
dropped the use of composite identities being behind a license check.
As a result, we can use `let_it_be` with the service account user
since the value of `composite_identity_enforced` will remain true.