Fix flaky diff note feature spec
Based on screenshots of failures, the spec is trying to hover on
line 22 of files/ruby/regex.rb to add a note on that specific
line but the hover and "add comment" button is showing up on a
different line which results to failure as it can't find the
"add comment" button for line 22.

We're scrolling too much in this spec though so this removes a
call to scroll and clicking on a line as it's also being done
in the `it` block for the same exact line.