Detail view renders legcy issue view for unsupported work item types
Service Desk issues and legacy incidents are not
fully migrated to the work items framework yet and
so should not render the work item detail view, even
if the user preference `use_work_items_view` is set
to `true`.

Because the user preference is hidden behind a
feature flag we don't consider this a feature change.

Changelog: other