Fix Amazon Q trigger service not triggering on new issuable
Previously `line_position_for_comment` was only added for the
`/q test` case, but when it was upstreamed it was always included.
The problem is that when the quick action is created from a new issue
or merge request, no note is available. Fix this by adding a safe
operator.

Also fix the error handling so that we can actually debug these
failures.