Log all AI Gateway responses related to Amazon Q
When the Amazon Q work was upstreamed, we dropped the use of
`Gitlab::Llm::AiGateway::Client#complete` in favor of direct HTTPS
POST requests. However, this dropped the logging that came with that
method.

Re-add this logging so we can track all responses from the AI Gateway.
All messages go to `llm.log`.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/510045

Changelog: added