Refactor sbom ingestion: extract VulnerabilityData struct to class
Context
--------------

As part of the [epic][0] to delete the
`vulnerability_finding_pipelines` table, we needed to migrate any
application code using that table to use new queries that did not
reference that table

As part of that effort, the struct being extracted was added in [MR
162713][1]

That table has been [dropped][2] for awhile now, and this
implementation seems stable.

This change
--------------

Now that these code paths are stable, the primary concern is
maintenance.

The code is currently covered by specs that test the
=IngestOccurrence= task, however there are no unit tests targeting the
query specifically. It would be nice to have this pulled out to a
class to allow for:

1. targeted specs to be written against any future bugs that might be
   uncovered
2. Being able to test a new version of the query with a FF in an
   easily isolated setup

This pulls the struct into its own and

1. adds unit tests
2. adds class documentation[^1]

[0]:https://gitlab.com/groups/gitlab-org/-/epics/11241
[1]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/162713
[2]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/170153

[^1]: We don't normally add class documentation. I added it here as it
      is a "sbom ingestion workflow" specific data object, and the
      documentation might help future readers grok that it is not a data
      object or concept used in the wider gitlab codebase