Create adapter service
Adds the EpicIssue Creation Adapter service which uses the WorkItem
service underneath, but on the surface has the same API as
the `EpicIssue::CreateService`.

For now this code is not used anywhere else to limit the changes
in this MR. It should be mostly a drop-in replacement though as
we're able to re-use the existing service specs.