Chore: fix vue3 tests crm
Additionally sets up the CRM apps to have a top level app.vue file
so that the router file can have a `component` property.
The component property is required in Vue router 4 and will throw
a warning if not set, therefore a wrapper app was necessary.

The new and edit routes are children routes, as they are a nested
router-view within the main view.