Ignore Elasticsearch database columns
We moved some Elasticsearch application settings to a new JSONB column
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/174172.

In that migration, we also dropped the original columns without following the
multi-milestone release process. This caused and incident:
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/19079.

We updated the last migration keeping the original columns
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/176987.

At the moment, we have the original columns and the JSONB columns with a jsob
accessor. When we update any setting, it get updated in both places.
Now, we will follow the process to drop columns:
https://docs.gitlab.com/ee/development/database/avoiding_downtime_in_migrations.html#dropping-columns