Update configured_at timestamp when no policies are changed
We use `configured_at` to determine which security orchestration
configurations are outdated, using scope `with_outdated_configuration`.

A cron job `CreateOrchestrationPolicyWorker` schedules a policy sync
based on that. However, if the repository is updated and the policies
don't change, we keep triggering the `Security::SyncScanPoliciesWorker`
every 10 minutes for such configuration.

This change fixes it to update `configured_at` even if the policies
don't change.