Documentation: clarify passing of artifacts
Currently the description of how artifacts are passed to jobs is under
a title "Prevent a job from fetching artifacts" ... which isn't the
first thing you look for when you're trying to understand how
artifacts are passed.  You can see a lot of the confusion in [1].

The best description of this is really the "dependencies:" tag
documentation, but similarly that isn't mentioned as the primary way
to pass artifacts.

Add a new second-level title "Using artifacts" which clearly points
you to the dependencies doc, with a shorter general explaination.  Add
the part on preventing the passing as a subsection of this.

[1] https://stackoverflow.com/questions/38140996/how-can-i-pass-gitlab-artifacts-to-another-stage

Co-authored-by: Lysanne Pinto <lpinto@gitlab.com>