BBS handle missing users when user mapping
In some cases the API returns no user details. When user mapping was
enabled, we would previously try to create an `Import::SourceUser` with
a `nil` `source_user_identifier`
https://gitlab.com/gitlab-org/gitlab/-/blob/72e83158cbee64324a4790dbac0777fc8fe9d494/lib/gitlab/bitbucket_server_import/user_finder.rb#L66
which led to an `ActiveRecord::RecordInvalid` error.

This change returns a `nil` user ID in this situation which is handled
by existing code in the same way as when user mapping is disabled.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/165855#note_2283797600
https://new-sentry.gitlab.net/organizations/gitlab/issues/1254442

Changelog: fixed