Drop unnecessary relative position spec expectation for Ruby 3.3
The previous spec assumed that when sorting an array with keys of
equal values, the order of the elements will always change (unstable
sort). However, it appears in Ruby 3.3 the output sometimes can be a
stable sort, so we shouldn't assume the order must always change. This
is a Ruby implementation detail.